pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "A statistics app"
authors = ["Max Mustermann <max.megamustermann@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-cors-headers = "^4.4.0"
djangorestframework = "^3.15.2"
pyjwt = "^2.9.0"
pytz = "^2024.2"
sqlparse = "^0.5.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
asgiref = "^3.8.1"
djangorestframework-simplejwt = "^5.3.1"
yfinance = "^0.2.44"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
